requirements.txt
asgiref==3.2.10
boto3==1.16.6
botocore==1.19.6
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.42.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-storages==1.10.1
Django-Verify-Email==0.0.3
gunicorn==20.0.4
jmespath==0.10.0
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.6
python3-openid==3.2.0
pytz==2020.1
requests-oauthlib==1.3.0
s3transfer==0.3.3
sqlparse==0.3.1
stripe==2.54.0
